# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FLAI API
  description: |
    Fully Local AI (FLAI) - API Documentation

    FLAI is a fully local personal assistant powered by artificial intelligence.

    ## Authentication
    Most endpoints require authentication via session cookie.
    Authenticate by POSTing to `/login` with credentials.

    ## Rate Limiting
    Login endpoint is rate limited: 5 requests per minute, 10 per hour.
    Send message endpoint: 15 per minute, 60 per hour.

    ## Feature Gates
    - Camera endpoints (`/api/cam/*`) require `CAMERA_ENABLED=true`
    - Debug endpoints (`/api/debug/*`, `/api/rag/*`) require `DEBUG_API_ENABLED=true`

    ## Error Responses
    All error responses return JSON with an `error` field:
    ```json
    {
      "error": "Error message"
    }
    ```
  version: 1.0.0
  contact:
    name: FLAI Support
    url: https://github.com/barval/flai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local development server
  - url: https://flai.example.com
    description: Production server
...
paths:
...
  /api/rag/search:
    post:
      tags:
        - Debug
      summary: Debug RAG search
      description: Direct RAG search returning raw chunks (requires DEBUG_API_ENABLED)
      parameters:
        - name: DEBUG_API_ENABLED
          in: query
          description: Requires DEBUG_API_ENABLED=true in .env
          schema:
            type: boolean
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                user_id:
                  type: string
                top_k:
                  type: integer
                  default: 80
      responses:
        200:
          description: RAG search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  user_id:
                    type: string
                  chunks_found:
                    type: integer
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/RAGChunk'
        400:
          description: Missing JSON body or query
        403:
          description: Debug API disabled
        503:
          description: RAG module not available
  /api/rag/stats:
    get:
      tags:
        - Debug
      summary: Debug RAG stats
      description: Get RAG module configuration (requires DEBUG_API_ENABLED)
      responses:
        200:
          description: RAG configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: boolean
                  top_k:
                    type: integer
                  chunk_size:
                    type: integer
                  chunk_overlap:
                    type: integer
                  chunk_strategy:
                    type: string
                  collection_name_prefix:
                    type: string
        403:
          description: Debug API disabled
        503:
          description: RAG module not available
...
  /api/debug/documents:
    get:
      tags:
        - Debug
      summary: Debug documents
      description: List all documents for a user (requires DEBUG_API_ENABLED)
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        200:
          description: Document list
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  documents_count:
                    type: integer
                  documents:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        filename:
                          type: string
                        file_path:
                          type: string
                        indexed_at:
                          type: string
                          format: date-time
                          nullable: true
                        file_size:
                          type: integer
        403:
          description: Debug API disabled
        500:
          description: Database error
  /api/debug/qdrant/{user_id}:
    get:
      tags:
        - Debug
      summary: Qdrant collection info
      description: Get Qdrant vector collection info for a user (requires DEBUG_API_ENABLED)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Collection info
          content:
            application/json:
              schema:
                type: object
                properties:
                  collection:
                    type: string
                  vectors_count:
                    type: integer
                  points_count:
                    type: integer
                  segments_count:
                    type: integer
        403:
          description: Debug API disabled
        500:
          description: Qdrant error
        503:
          description: RAG not available
  /api/debug/qdrant/{user_id}/search:
    post:
      tags:
        - Debug
      summary: Qdrant direct search
      description: Direct Qdrant vector search without RAG processing (requires DEBUG_API_ENABLED)
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                limit:
                  type: integer
                  default: 10
      responses:
        200:
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  limit:
                    type: integer
                  found:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        doc_id:
                          type: string
                        score:
                          type: number
                        text:
                          type: string
        400:
          description: Missing JSON or query
        403:
          description: Debug API disabled
        500:
          description: Search error
        503:
          description: RAG not available
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    RAGChunk:
      type: object
      properties:
        index:
          type: integer
        doc_id:
          type: string
        text:
          type: string
        score:
          type: number
        metadata:
          type: object